+++ src/components/ui/category-swiper.tsx
