+++ src/components/ui/ticket-card.tsx
