+++ src/features/chatbot/components/help-button.tsx
